Rename instructions to `request` and `http-request`
