+++ lib/runner/extensions/event.command.js
@@ -258,7 +258,7 @@ module.exports = {
 
                         cursorWithSource._source = 'script'; // @todo - get script type from the sandbox
 
-                        this.immediate('request', {
+                        this.immediate('httpRequest', {
                             item: tempItem,
                             globals: payload.context.globals,
                             environment: payload.context.environment,
