+++ lib/runner/extensions/item.command.js
@@ -83,7 +83,7 @@ module.exports = {
                         });
                     }
 
-                    this.queue('requestWrapper', {
+                    this.queue('request', {
                         item: item,
                         globals: ctxTemplate.globals,
                         environment: ctxTemplate.environment,
