+++ lib/runner/replay-controller.js
@@ -57,7 +57,7 @@ _.assign(ReplayController.prototype, /** @lends ReplayController.prototype */{
             replayState: context.replayState
         };
 
-        this.run.immediate('request', payload)
+        this.run.immediate('httpRequest', payload)
             .done(function (response) {
                 success(null, response);
             })
