+++ lib/runner/run.js
@@ -191,8 +191,8 @@ _.assign(Run, {
 Run.commands = _.transform({
     'control.command': require('./extensions/control.command'),
     'event.command': require('./extensions/event.command'),
+    'httpRequest.command': require('./extensions/http-request.command'),
     'request.command': require('./extensions/request.command'),
-    'requestWrapper.command': require('./extensions/request-wrapper.command'),
     'waterfall.command': require('./extensions/waterfall.command'),
     'item.command': require('./extensions/item.command'),
     'delay.command': require('./extensions/delay.command')
